Reporte Clientes Trimestral/requirements.txt
streamlit>=1.30,<2.0
python-pptx>=0.6.21
feedparser>=6.0.11
openai>=1.0.0
